chore: update and consolidate parseCoins into cosmjs code
